requirements.txt
google-api-python-client>=2.140.0
google-auth-oauthlib>=1.2.0
google-auth>=2.29.0
google-genai>=1.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
